Initial Code for Init and Update Functions(RTOS NOT IMPLEMENTED YET)
Designed functions to account for wheel speeds of the rear and front wheels to determine slip ratio.
NOT YET COMPLETED
Couple To-Dos:
Acclimate it with FreeRTOS for a specific timeframe.
Tune the PID Control values accordingly (this may have to be a testing phase thing)
Do note that this code is assuming that the wheel speed sensors are not provided in radians, but if they are ( which lowkey they most likely are), then we gotta convert them by just multiplying by the radius of the wheels im assuming
Currently unsure how the outputs might come out. One thing I would like to try is designing our own like fake simulation using w and d as accelerator and break and maybe finding a way to do smth with that (not too sure tho just a  thought type of idea ykyk)