Add sensor read / integrate
Converted Sensor outputs (0-5V) into readable values. Initialized analog pins 0-2 temporarily and max resolution of teensy set at 10 bits in this, can be scaled up to 12 i believe.

TODO: Unsure of stack size, could be bigger im not sure if 128 words is going to be enough.
change max values for shock travel and torque sensor. Maybe set it so that if it exceeds the max values, then send a warning?
Not entirely sure what pins and how pins are mapped to what so maybe double check this